# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ThemeMetrics API
  description: API für Shopify Theme Performance Analyse
  version: 1.0.0
  contact:
    name: ThemeMetrics Support
    url: https://thememetrics.de
servers:
  - url: https://thememetrics.de/api
    description: Production
  - url: http://localhost:3000/api
    description: Development
security:
  - sessionToken: []
  - cookieAuth: []
...
paths:
  /themes/analyze:
    post:
      tags: [Analysis]
      summary: Theme analysieren
      description: Analysiert das aktive Theme eines Shopify Stores
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shop:
                  type: string
                  example: example.myshopify.com
              required: [shop]
      responses:
        '200':
          description: Analyse erfolgreich
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/LimitReached'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /webhooks:
    post:
      tags: [Webhooks]
      summary: Shopify Webhook Empfänger
      security: []
      parameters:
        - name: X-Shopify-Topic
          in: header
          required: true
          schema:
            type: string
        - name: X-Shopify-Shop-Domain
          in: header
          required: true
          schema:
            type: string
        - name: X-Shopify-Hmac-Sha256
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook verarbeitet
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: Shopify Session Token
    cookieAuth:
      type: apiKey
      in: cookie
      name: shop_session
  schemas:
    AnalysisResult:
      type: object
      properties:
        theme:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            role:
              type: string
            analyzedAt:
              type: string
              format: date-time
        analysis:
          type: object
          properties:
            id:
              type: integer
            totalSections:
              type: integer
            sections:
              type: array
              items:
                $ref: '#/components/schemas/Section'
            highImpactCount:
              type: integer
        score:
          $ref: '#/components/schemas/ScoreBreakdown'
        plan:
          type: object
          properties:
            current:
              type: string
            showSectionDetails:
              type: boolean
            maxRecommendations:
              type: integer
    ScoreBreakdown:
      type: object
      properties:
        overall:
          type: integer
          minimum: 0
          maximum: 100
        speed:
          type: object
          properties:
            score:
              type: integer
            coreWebVitals:
              type: integer
            sectionLoad:
              type: integer
            details:
              $ref: '#/components/schemas/CoreWebVitals'
        quality:
          type: object
          properties:
            score:
              type: integer
            liquidQuality:
              type: integer
            bestPractices:
              type: integer
            architecture:
              type: integer
        conversion:
          type: object
          properties:
            score:
              type: integer
            ecommerce:
              type: integer
            mobile:
              type: integer
            estimatedMonthlyLoss:
              type: number
        hasRealData:
          type: boolean
    CoreWebVitals:
      type: object
      properties:
        lcp:
          type: number
          description: Largest Contentful Paint (ms)
        cls:
          type: number
          description: Cumulative Layout Shift
        fcp:
          type: number
          description: First Contentful Paint (ms)
        tbt:
          type: number
          description: Total Blocking Time (ms)
    Section:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        category:
          type: string
        performanceScore:
          type: integer
        performanceImpact:
          type: integer
        recommendations:
          type: array
          items:
            type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        upgradeRequired:
          type: boolean
  responses:
    Unauthorized:
      description: Nicht authentifiziert
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: not_authenticated
            message: Session fehlt oder ungültig
    LimitReached:
      description: Plan-Limit erreicht
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: limit_reached
            message: Monatliches Limit erreicht
            upgradeRequired: true
    RateLimited:
      description: Rate Limit erreicht
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: rate_limited
            message: Zu viele Requests